Add a setting to set if report a 401 error on invalid pattern

+++ caselink/serializers.py
@@ -2,24 +2,28 @@ from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 from django.utils.translation import ugettext as _
 from django.core.exceptions import ObjectDoesNotExist
+from django.conf import settings
 from .models import *
 
 
 class LinkageSerializer(serializers.ModelSerializer):
     def validate_autocase_pattern(self, data):
-        # TODO: avoid creating new instead of deleting old
-        all_data = self.initial_data
-        linkage = Linkage.objects.filter(
-            autocase_pattern=data,
-            workitem=all_data['workitem']
-        )
-        if len(linkage) == 1:
-            linkage[0].delete()
+        if settings.CASELINK['401_ON_INVALID_PATTERN']:
+            # TODO: avoid creating new instead of deleting old
+            all_data = self.initial_data
+            linkage = Linkage.objects.filter(
+                autocase_pattern=data,
+                workitem=all_data['workitem']
+            )
+            if len(linkage) == 1:
+                linkage[0].delete()
 
-        for case in AutoCase.objects.all():
-            if test_pattern_match(data, case.id):
-                return data
-        raise serializers.ValidationError("Pattern Invalid.")
+            for case in AutoCase.objects.all():
+                if test_pattern_match(data, case.id):
+                    return data
+            raise serializers.ValidationError("Pattern Invalid.")
+        else:
+            return data
 
     class Meta:
         fields = '__all__'
